Adicionar status http organizacao

+++ collecta/src/main/java/collectiva/org/collecta/service/OrganizacaoService.java
@@ -23,11 +23,17 @@ public class OrganizacaoService {
 
     public ResponseEntity<List<Organizacao>> buscarTodasOrganizacoes() {
         List<Organizacao> organizacao = organizacaoRepository.findAll();
+        if (organizacao.isEmpty()){
+            return ResponseEntity.noContent().build();
+        }
         return ResponseEntity.ok().body(organizacao);
     }
 
     public ResponseEntity<Optional<Organizacao>> buscarOrganizacaoPorId(UUID id) {
         Optional<Organizacao> organizacao = organizacaoRepository.findById(id);
+        if (organizacao.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok().body(organizacao);
     }
 
@@ -52,6 +58,9 @@ public class OrganizacaoService {
     }
 
     public ResponseEntity<Void> deletarOrganizacao(UUID id) {
+        if (organizacaoRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         organizacaoRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
